requirements.txt
jinja2==3.1.6
livereload==2.7.1
more-itertools==10.7.0
python-dotenv==1.1.0